Adding an auto-generated pacing calendar option. To show the calendar, I’ve added a flag on the unit edit page called “show calendar”. When checked, the calendar will appear on /<course>/<unit> and /<course>/<unit>/<glance>
By default, lessons within the same week are evenly spaced on the calendar. An additional field called “pacing weight” has been added to lessons to override this. Increasing the pacing weight will make the lesson take up more space on the calendar.

I also added a multi-lesson edit view to the bottom of unit edit page to make it easier to change weighting, as well as keywords and other flags.