Fix documentation url routing
At some point recently the documentation routes stopped working. This was likely caused by some of the changes I made to move map levels into docs.code.org. This corrects for that issue as well as fixing a long standing bug that prevented the admin "view on site" button from using the correct local route for ides and blocks