Add ability to specify whether or not a model should be redacted
This is needed specifically for Resources; one of the fields we are
translating for them is a raw URL, which gets parsed as a markdown URL
element by the redaction process, resulting in broken links.

More generally, it's not good for the system to assume that everything
needs to be redacted, or to characterize "redaction" as being synonymous
with "preparing to uplod to crowdin".

This adds the ability for models to specify whether or not they should
be redacted, and adds awareness to the i18n sync process to take
advantage of that. It also renames the directory into which we put
strings to be uploaded to crowdin to make it more clear that redaction
is merely a _part_ of the process.

TODO: Make it possible to specify not just a subset of models to
redact, but a subset of fields on a model.