Only restore redacted content
Follow-up to https://github.com/mrjoshida/curriculumbuilder/pull/85

The goal of that PR was to only redact content we know we want to be
redacted, to prevent applying markdown formatting to things we don't
want to treat as markdown (like Resource URLs). Unfortunately, even
though we're only _redacting_ the things we know we want to process,
we're still _restoring_ everything, which applies the same formatting as
redaction.

Extending the requirement to both redaction and restoration should
actually fix the problem.

Also broke the `handle` method up into individual helper methods, since
it was getting too long.