Use strings for i18n keys
Since our i18n values are communicated through JSON, we always want to
use strings. Using non-string values will work fine for source string
generation, since the keys will be converted to strings in the
serialization process; but it will fail to find the translations since
the string keys won't be converted back to integers in the
deserialization process.