Invoke CodeMirror later on map level template
We're seeing an issue where CodeMirror is applying incorrect spacing to
HTML snippets only when the page is rendered within an iframe on a Code
Studio map level.  This seems to be due to a timing issue - the
CodeMirror handler is being invoked before the document in the iframe
has fully loaded, leading to some incorrect calculations.

Here I'm suggesting we use jQuery's `$(window).on("load", fn)` instead
of `$(fn)` / `$(document).ready(fn)` which explicitly waits for the
whole page (including iframes) to be ready, not just the DOM.

Read more: https://learn.jquery.com/using-jquery-core/document-ready/