Migrate PDF generation off of wkhtmltopdf onto Puppeteer
Wkhtmltopdf was dying on some of our larger pages. Investegation revealed that
this was due a problem in the C library, that we would need to improve logging
in the C library to get to the bototm of.

Rather than do that, we migrate to Puppeteer. Puppeteer does not run into the
problems we were seeing with Wkhtmltopdf. Puppeteer is maintained by Google,
whereas Wkhtmltopdf is maintained by two developers in their spare time.
Puppeteer has clearer error reporting and other documentation than Wkhtmltopdf,
and we expect will continue to be better supported in the future.

Thanks to Winter for researching Puppeteer for this usecase