Fixing divide-by-zero error in publish_i18n
We only publish changes for a language if there have been changes downloaded. There is some logging code which divide by the number of languages with changes. If no languages have changed, then the logging code ends up dividing by 0. The fix is to skip the whole I18n publish step if we don't have any changes since it won't be publishing anything.